Updating latest dependencies.
Includes MongoDB, Express, Socket.IO, and all the other little ones.
Fixes #52.
Basically, the package.json was too loose, allowing newer
versions of packages which broke backwards compatability.
So, it's a bit more strict now, and the code has been
updated to take advantage of all the current
versions of the packages anyway.